refactor(one): consolidate babel plugin chain into one/babel-preset
