+++ packages/one/types/babel-preset/index.d.ts
@@ -1,4 +1,4 @@
-import type { TransformOptions } from '@babel/core';
+import type { PluginItem, TransformOptions } from '@babel/core';
 type BabelConfigAPI = {
     cache?: ((forever?: boolean) => void) & {
         forever?: () => void;
@@ -27,28 +27,42 @@ export type OneBabelPresetOptions = {
     };
     /** Whether to include `babel-preset-expo` as the base preset. Defaults to true. */
     includeExpoPreset?: boolean;
+    /**
+     * Whether to include `@vxrn/vite-plugin-metro/babel-plugins/import-meta-env-plugin`.
+     * Defaults to true. The Vite-driven Metro server injects this separately via
+     * `patchMetroServerWithViteConfigAndMetroPluginOptions` using the user's Vite
+     * `define` config — so the Vite path passes `false`. Re-applying is idempotent.
+     */
+    includeImportMetaEnv?: boolean;
 };
 /**
  * Standalone babel preset that drops the same plugin chain that the
  * Vite-driven Metro path applies into any `babel.config.{cjs,js,mjs}` file.
- *
- * Use this from a project's `babel.config.cjs` to make `expo export`,
- * `eas update`, and other Metro-direct workflows produce byte-equivalent
- * bundles to what `vxrn`/`one dev`/`one build` produce internally.
  *
  * @example
  * ```js
  * // babel.config.cjs
  * module.exports = require('one/babel-preset')
  * ```
- *
- * @example
- * ```js
- * // babel.config.cjs (with options)
- * module.exports = (api) =>
- *   require('one/babel-preset')(api, { routerRoot: 'src/routes' })
- * ```
  */
 export default function oneBabelPreset(api: BabelConfigAPI, options?: OneBabelPresetOptions): TransformOptions;
+export type BuildOneBabelPluginsOptions = {
+    projectRoot: string;
+    relativeRouterRoot: string;
+    ignoredRouteFiles?: Array<`**/*${string}`>;
+    linking?: unknown;
+    setupFile?: string | {
+        native?: string;
+        ios?: string;
+        android?: string;
+    };
+    includeImportMetaEnv?: boolean;
+};
+/**
+ * The plugin chain shared between the Vite-driven Metro path
+ * (`getViteMetroPluginOptions`) and the standalone preset above.
+ */
+export declare function buildOneBabelPlugins({ projectRoot, relativeRouterRoot, ignoredRouteFiles, linking, setupFile, includeImportMetaEnv, }: BuildOneBabelPluginsOptions): PluginItem[];
+export declare function buildRouterRequireContextRegexString(ignoredRouteFiles?: Array<`**/*${string}`>): string;
 export {};
 //# sourceMappingURL=index.d.ts.map
