--- packages/one/types/metro-config/buildOneBabelPlugins.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=buildOneBabelPlugins.test.d.ts.map
