--- packages/one/types/metro-config/getViteMetroPluginOptions.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=getViteMetroPluginOptions.test.d.ts.map
